+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 + Add Bank Account Transactions service
 + Add Bank Account Transactions controller + DTO
 + Add Docker for deploy
++ Add Bank Account Transactions service POST
 
 ## 0.1.0
 + First commit